move postUserStatus from confirmation page to viewUser page
