+++ app/pages/viewuser/changeUserStatusConfirmation/controller.js
@@ -1,4 +1,4 @@
-import { getData, postData } from '../../../apiProvider';
+import { getData } from '../../../apiProvider';
 import { getContext } from './contextCreator';
 import { logger } from '../../../logger';
 
@@ -6,6 +6,7 @@ export const getUserStatusContext = async ({
   userId, organisationId, accessToken, status,
 }) => {
   const user = await getData({ endpointLocator: 'getUserById', options: { userId }, accessToken });
+  user.userId = userId;
   if (user) {
     logger.info(`User ${user.userId}: ${user.name} found`);
     return getContext({
@@ -13,15 +14,4 @@ export const getUserStatusContext = async ({
     });
   }
   throw new Error(`No user data returned for id: ${userId}`);
-};
-
-export const postUserStatus = async ({ userId, accessToken, status }) => {
-  try {
-    await postData({ endpointLocator: 'postUserStatus', options: { userId, status }, accessToken });
-
-    logger.info(`User ${userId} status updated to: ${status}`);
-    return { success: true };
-  } catch (err) {
-    throw new Error(`Unable to update status for user id: ${userId}`);
-  }
 };
