+++ app/pages/viewuser/changeUserStatusConfirmation/controller.test.js
@@ -1,10 +1,9 @@
-import { getUserStatusContext, postUserStatus } from './controller';
+import { getUserStatusContext } from './controller';
 import * as apiProvider from '../../../apiProvider';
 import * as contextCreator from './contextCreator';
 
 jest.mock('../../../apiProvider', () => ({
   getData: jest.fn(),
-  postData: jest.fn(),
 }));
 
 jest.mock('./contextCreator', () => ({
@@ -12,6 +11,7 @@ jest.mock('./contextCreator', () => ({
 }));
 
 const mockUserData = {
+  userId: 'user1',
   name: 'Some',
   phonenumber: '07777777777',
   emailaddress: 'some@body.com',
@@ -30,12 +30,12 @@ describe('user status controller', () => {
         apiProvider.getData
           .mockResolvedValueOnce(mockUserData);
         await getUserStatusContext({
-          userId: 1, organisationId: 2, accessToken: 'access_token', status: 'enable',
+          userId: 'user1', organisationId: 'org1', accessToken: 'access_token', status: 'enable',
         });
         expect(apiProvider.getData.mock.calls.length).toEqual(1);
         expect(apiProvider.getData).toHaveBeenCalledWith({
           endpointLocator: 'getUserById',
-          options: { userId: 1 },
+          options: { userId: 'user1' },
           accessToken: 'access_token',
         });
       });
@@ -47,12 +47,12 @@ describe('user status controller', () => {
           .mockResolvedValueOnce();
 
         await getUserStatusContext({
-          userId: 1, organisationId: 2, accessToken: 'access_token', status: 'enable',
+          userId: 'user1', organisationId: 'org1', accessToken: 'access_token', status: 'enable',
         });
 
         expect(contextCreator.getContext.mock.calls.length).toEqual(1);
         expect(contextCreator.getContext).toHaveBeenCalledWith({
-          user: { ...mockUserData }, status: 'enable', organisationId: 2,
+          user: { ...mockUserData }, status: 'enable', organisationId: 'org1',
         });
       });
     });
@@ -67,12 +67,12 @@ describe('user status controller', () => {
         apiProvider.getData
           .mockResolvedValueOnce(mockUserData);
         await getUserStatusContext({
-          userId: 1, organisationId: 2, accessToken: 'access_token', status: 'disable',
+          userId: 'user1', organisationId: 'org1', accessToken: 'access_token', status: 'disable',
         });
         expect(apiProvider.getData.mock.calls.length).toEqual(1);
         expect(apiProvider.getData).toHaveBeenCalledWith({
           endpointLocator: 'getUserById',
-          options: { userId: 1 },
+          options: { userId: 'user1' },
           accessToken: 'access_token',
         });
       });
@@ -84,69 +84,26 @@ describe('user status controller', () => {
           .mockResolvedValueOnce();
 
         await getUserStatusContext({
-          userId: 1, organisationId: 2, accessToken: 'access_token', status: 'disable',
+          userId: 'user1', organisationId: 'org1', accessToken: 'access_token', status: 'disable',
         });
 
         expect(contextCreator.getContext.mock.calls.length).toEqual(1);
         expect(contextCreator.getContext).toHaveBeenCalledWith({
-          user: { ...mockUserData }, status: 'disable', organisationId: 2,
+          user: { ...mockUserData }, status: 'disable', organisationId: 'org1',
         });
       });
     });
 
     it('should throw an error when no user data is returned from the apiProvider', async () => {
       apiProvider.getData
-        .mockResolvedValueOnce();
+        .mockResolvedValueOnce({});
 
       try {
         await getUserStatusContext({
-          userId: 1, organisationId: 2, accessToken: 'access_token', status: 'disable',
+          userId: 'user1', organisationId: 'org1', accessToken: 'access_token', status: 'disable',
         });
       } catch (err) {
         expect(err).toEqual(new Error('No user data returned for id: 1'));
-      }
-    });
-  });
-
-  describe('postAddUser', () => {
-    afterEach(() => {
-      apiProvider.postData.mockReset();
-    });
-
-    it('should call postUserStatus once with the correct params', async () => {
-      apiProvider.postData
-        .mockResolvedValueOnce({});
-
-      await postUserStatus({ userId: 1, accessToken: 'access_token', status: 'enable' });
-
-      expect(apiProvider.postData.mock.calls.length).toEqual(1);
-      expect(apiProvider.postData).toHaveBeenCalledWith({
-        endpointLocator: 'postUserStatus',
-        options: {
-          userId: 1,
-          status: 'enable',
-        },
-        accessToken: 'access_token',
-      });
-    });
-
-    it('should return true if api request is successful', async () => {
-      apiProvider.postData
-        .mockResolvedValueOnce({});
-
-      const response = await postUserStatus({ userId: 1, accessToken: 'access_token', status: 'enable' });
-
-      expect(response.success).toEqual(true);
-    });
-
-    it('should throw an error if api request is unsuccessful', async () => {
-      apiProvider.postData
-        .mockRejectedValueOnce({ response: { status: 500 } });
-
-      try {
-        await postUserStatus({ userId: 1, accessToken: 'access_token', status: 'enable' });
-      } catch (err) {
-        expect(err).toEqual(new Error('Unable to update status for user id: 1'));
       }
     });
   });
