+++ app/pages/viewuser/controller.js
@@ -1,4 +1,4 @@
-import { getData } from '../../apiProvider';
+import { getData, postData } from '../../apiProvider';
 import { getContext } from './contextCreator';
 import { logger } from '../../logger';
 
@@ -15,4 +15,15 @@ export const getViewUserContext = async ({ organisationId, userId, accessToken }
   }
 
   throw new Error(`No user data returned for id: ${userId}`);
+};
+
+export const postUserStatus = async ({ userId, accessToken, status }) => {
+  try {
+    await postData({ endpointLocator: 'postUserStatus', options: { userId, status }, accessToken });
+
+    logger.info(`User ${userId} status updated to: ${status}`);
+    return { success: true };
+  } catch (err) {
+    throw new Error(`Unable to update status for user id: ${userId}`);
+  }
 };
