+++ app/routes.js
@@ -7,8 +7,8 @@ import { getOrgDashboardContext } from './pages/dashboard/controller';
 import { getAddUserContext, getAddUserPageErrorContext, postAddUser } from './pages/adduser/controller';
 import includesContext from './includes/manifest.json';
 import { getAddUserConfirmationContext } from './pages/adduser/confirmation/controller';
-import { getViewUserContext } from './pages/viewuser/controller';
-import { getUserStatusContext, postUserStatus } from './pages/viewuser/changeUserStatusConfirmation/controller';
+import { getViewUserContext, postUserStatus } from './pages/viewuser/controller';
+import { getUserStatusContext } from './pages/viewuser/changeUserStatusConfirmation/controller';
 import { getEditOrgContext, putUpdateOrganisation } from './pages/editorg/controller';
 import { getEditOrgConfirmationContext } from './pages/editorg/confirmation/controller';
 import config from './config';
